__tests__/main.test.ts
import {expect, test} from '@jest/globals'
import {process} from '../src/main'

test('testParsing', async () => {
  console.log('dummy')
  //const bom = parseSbomFile('data/sample.json')
  //expect(bom).not.toBeNull()
})

test('runAll', async () => {
  await process('data/valid-bom-1.4.json')
})

/*
test('throws invalid number', async () => {
  const input = parseInt('foo', 10)
  await expect(wait(input)).rejects.toThrow('milliseconds not a number')
})

test('wait 500 ms', async () => {
  const start = new Date()
  await wait(500)
  const end = new Date()
  var delta = Math.abs(end.getTime() - start.getTime())
  expect(delta).toBeGreaterThan(450)
})

// shows how the runner will run a javascript action with env / stdout protocol
test('test runs', () => {
  process.env['INPUT_MILLISECONDS'] = '500'
  const np = process.execPath
  const ip = path.join(__dirname, '..', 'lib', 'main.js')
  const options: cp.ExecFileSyncOptions = {
    env: process.env
  }
  console.log(cp.execFileSync(np, [ip], options).toString())
})
*/
